Move everything into a class
Instead of keeping everything in one object with global state, we can
move it all into a class an export an instantiated object of it. This
way, for the general case, all uses of this default instantiated will
match the settings for the top level version.

However, a new method is provided which offers a way to instantiate a
new version. This allows for the case where different protocol versions
might need to be loaded.